requirements.txt
google-cloud-spanner
faker
